in dist script run mvn build and check exit status

+++ dist_template/bin/dist.sh
@@ -8,7 +8,14 @@ JARS="graph-builder"
 WINSTONE_URL="http://sourceforge.net/projects/winstone/files/winstone/v0.9.10/winstone-0.9.10.jar/download"
 
 cd $BASE
-#mvn clean package
+
+mvn clean package
+# check for build failure
+if [ $? -ne 0 ]; then
+   echo "MAVEN BUILD FAILED"
+   exit 1
+fi
+
 rm -rf $BASE/dist
 cp -R $BASE/dist_template $BASE/dist
 mkdir -p $BASE/dist/webapps
